refactor: extract shared pipeline module from main.rs
Move load_and_price, parse_with_cache, resolve_providers, and helpers
into src/pipeline.rs. Add to lib.rs for library consumers. Update
mcp.rs to use crate::pipeline::load_and_price instead of
super::load_and_price. Reduces main.rs by ~170 lines and eliminates
the dependency of mcp on the binary entry point.
